+++ providers/rippling/gen/main.go
@@ -4,8 +4,8 @@
 package main
 
 import (
-	"go.mondoo.com/mql/v13/providers-sdk/v1/plugin/gen"
-	"go.mondoo.com/mql/v13/providers/rippling/config"
+	"go.mondoo.com/mql/providers-sdk/v1/plugin/gen"
+	"go.mondoo.com/mql/providers/rippling/config"
 )
 
 func main() {
