+++ providers/rippling/resources/rippling.lr
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: BUSL-1.1
 
 option provider = "go.mondoo.com/mql/providers/rippling"
-option go_package = "go.mondoo.com/mql/v13/providers/rippling/resources"
+option go_package = "go.mondoo.com/mql/providers/rippling/resources"
 
 // Rippling
 //
